FE-43🏗️ 상세페이지 Layout 구조개선
