+++ src/apis/epigram.ts
@@ -6,7 +6,7 @@ const BASE_URL = 'https://fe-project-epigram-api.vercel.app/5-9';
 const getEpigram = async (request: GetEpigramRequestType): Promise<GetEpigramResponseType> => {
   const { id } = request;
   const TOKEN =
-    'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MjIsInRlYW1JZCI6IjUtOSIsInNjb3BlIjoiYWNjZXNzIiwiaWF0IjoxNzIwNzUxNTM2LCJleHAiOjE3MjA3NTMzMzYsImlzcyI6InNwLWVwaWdyYW0ifQ.7TTCMekxh9_J9O0jPAgIlZmYdUMSGGVf7qkiYJqM9rk';
+    'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MjIsInRlYW1JZCI6IjUtOSIsInNjb3BlIjoiYWNjZXNzIiwiaWF0IjoxNzIwNzUzNjYwLCJleHAiOjE3MjA3NTU0NjAsImlzcyI6InNwLWVwaWdyYW0ifQ.HqHWi_Uoj37aZzC_43d2lh5-VR2b4mXZWO--905fCQQ';
 
   try {
     const response = await axios.get(`${BASE_URL}/epigrams/${id}`, {
